Added some unsigned type conversions to remove warnings.

+++ iobuf/ibuf_refill.c
@@ -27,7 +27,7 @@ int ibuf_refill(ibuf* in)
   if(io->buflen < io->bufsize) {
     if (io->timeout && !iobuf_timeout(io, 0)) return 0;
     rd = read(io->fd, io->buffer+io->buflen, io->bufsize-io->buflen);
-    if(rd == -1)
+    if(rd == (unsigned)-1)
       IOBUF_SET_ERROR(io);
     else if(rd == 0)
       io->flags |= IOBUF_EOF;
